+++ plugins/dates/index.d.ts
@@ -1,6 +1,21 @@
 import nlp from 'compromise'
 type View = ReturnType<typeof nlp>
 
+export interface dateOptions {
+  /** the default timezone is 'ETC/UTC' */
+  timezone?: string,
+  /** the implicit, or reference moment for 'now'*/
+  today: '2020-02-20',
+  /** the implied duration to use for 'after june 2nd' */
+  punt: { weeks: 2 },
+  /** the default beginning of a day, like '8:00am'*/
+  dayStart: string,
+  /** the default beginning of a day, like '5:00pm'*/
+  dayEnd: string,
+  /** whether to assume british-format dates, when unclear*/
+  dmy: boolean
+}
+
 interface DateView extends View {
   /** convert parsed dates to a date format */
   format(fmt: string): View
@@ -17,9 +32,9 @@ interface TimeView extends View {
 
 export interface DatesMethods {
   /** match all date-phrases */
-  dates(): DateView
+  dates(opts?: dateOptions): DateView
   /** match time-of-day phrases */
-  times(): TimeView
+  times(opts?: dateOptions): TimeView
   /** match lengths of time, like '2 weeks' */
   durations(): View
 }