+++ src/2-two/postTagger/model/nouns/places.js
@@ -25,6 +25,13 @@ export default [
     tag: 'Address',
     reason: 'address-st',
   },
+  // port dover
+  { match: '(port|mount|mt) #ProperName', tag: 'Place', reason: 'port-name' },
+  // generic 'oak ridge' names
+  // { match: '(oak|maple|spruce|pine|cedar|willow|green|sunset|sunrise) #Place', tag: 'Place', reason: 'tree-name' },
+  // generic 'sunset view' names
+  // { match: '() #Place', tag: 'Place', reason: 'tree-name' },
+
   // Sports Arenas and Complexs
   // {
   //   match: