+++ src/3-three/misc/plugin.js
@@ -3,6 +3,7 @@ import addParentheses from './parentheses/index.js'
 import addPossessives from './possessives/index.js'
 import addQuotations from './quotations/index.js'
 import addSelections from './selections/index.js'
+import addSlashes from './slashes/index.js'
 
 export default {
   api: function (View) {
@@ -11,5 +12,6 @@ export default {
     addPossessives(View)
     addQuotations(View)
     addSelections(View)
+    addSlashes(View)
   }
 }